requirements.txt
fastapi
fastapi-cache
uvicorn
redis
python-dotenv
sqlalchemy